requirements.txt
# Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0

# Image Processing
Pillow==10.1.0
numpy==1.26.2
scikit-image==0.22.0

# Async Tasks
celery==5.3.4
redis==5.0.1
flower==2.0.1

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1

# Caching & Rate Limiting
slowapi==0.1.9
python-jose[cryptography]==3.3.0

# Monitoring
prometheus-client==0.19.0

# Utilities
python-dotenv==1.0.0
aiofiles==23.2.1
httpx==0.25.2

# Code Quality
black==23.12.1
isort==5.13.2
flake8==7.0.0
mypy==1.7.1
pylint==3.0.3
ruff==0.1.9
pre-commit==3.6.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0

